requirements.txt
## Setup py requirements

#python -m spacy download en_core_web_sm
### On MacBook
#bash: /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
# bash: echo 'eval "$(/opt/homebrew/bin/brew shellenv)"' >> ~/.zprofile
#bash: eval "$(/opt/homebrew/bin/brew shellenv)"
#bash: arch -arm64 brew install pkg-config hdf5
#bash: unname -m
#then bash: tensorflow-macos==2.9


#python==3.8.20
requests==2.25.1
#rasa==3.6.21
#rasa-sdk==3.6.2

#spacy==3.0.6
#spacy-legacy==3.0.12
scikit-learn==1.0.2
numpy==1.23.5
#numpy==1.26.0

setuptools==70.3.0
pydantic>=1.7.1,<1.8.0
protobuf==4.23.3
absl-py==1.4
requests==2.32.4

#sys_platform == "darwin"
#platform_machine == "arm64"
# Database
sqlalchemy==1.4.8
psycopg2-binary>=2.9,<2.10
redis>=4.0

# YAML validatorS
pykwalify>=1.7.0,<2.0
ruamel.yaml
ujson # for json file format

# Optional (MacBook ARM support)
tensorflow-metal==0.7.0
#tensorflow-metal==1.2.0
tensorflow-macos==2.12.0
#tensorflow-macos==2.16.1
#tensorflow-metal==0.6.0

coloredlogs>=15.0,<16.0
aiohttp==3.9.0
flask-api==3.1


# UI specific frameworks
fbmessenger>=6.0.0,<6.1.0 # facebook
slackclient>=2.0.0,<3.0.0 # slack
#twilio>=6.26,<6.51  #twilio
#pyTelegramBotAPI>=3.7.3,<4.0.0 #telegram
#mattermostwrapper>=2.2,<2.3 #Mattermost / RocketChat
#rocketchat_API>=0.6.31,<1.17.0 # Mattermost / RocketChat
#webexteamssdk>=1.1.1,<1.7.0 # Webex

##For viewing rasa version
#bash: sudo rm -rf /Library/Developer/CommandLineTools
#bash: xcode-select --install
